Gns1719

+++ BE/pet_noseprint_id_service/src/main/java/com/example/pet_noseprint_id/pet/repository/PetRepository.java
@@ -7,6 +7,6 @@ import java.util.List;
 
 
 public interface PetRepository extends CrudRepository<Pet, Long> {
-    List<Pet> findByUserKey(Integer userKey);
-    boolean existsByUserKey(Integer userKey);
+    List<Pet> findByUserKey(Long userKey);
+    boolean existsByUserKey(Long userKey);
 }
